% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flights-package.R
\docType{package}
\name{flights-package}
\alias{flights}
\alias{flights-package}
\title{flights: Build a Package to Analysis Penguins Data}
\description{
Use Rmarkdown First method to build your package. Start your package with documentation. Everything can be set from a Rmarkdown file in your project.
}
\author{
\strong{Maintainer}: Bowen Qin \email{qinbowen181@163.com}

}
\keyword{internal}
