% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flights-package.r
\docType{package}
\name{flights}
\alias{flights}
\alias{flights-package}
\title{Extract Information on Flights}
\description{
Use this package to call functions that can get information on flights that isn't available in the `nycflights13` package.
}
\note{
This package was created for STAT547.
}
