% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel.R
\name{fuel}
\alias{fuel}
\title{Fuel used each flight}
\usage{
fuel(distance_km)
}
\arguments{
\item{distance_km}{The distance in kilometres.}
}
\value{
The fuel in litres used for the flight based on the flight path \code{distance_km}.
}
\description{
This function calculates the fuel used each flight based on the distance of the flight.
}
\details{
This function calculates the fuel used for each flight based on the distance (kilometres) of the flight and the assumption that the plane uses appromiately 12 litres of fuel per kilometre (1,200 litres/ 100 km).
The average fuel consumption was based on one of the most common commercial planes, the Boeing 747.
This function would be very helpful when using the package `nycflights13`.
}
\examples{
fuel(1)
fuel(250)
}
