% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestokm.R
\name{milestokm}
\alias{milestokm}
\title{Change from miles to kilometres}
\usage{
milestokm(mi)
}
\arguments{
\item{mi}{The distance in miles.}
}
\value{
The distance in kilometres of \code{mi}.
}
\description{
This function just converts the distance from miles to kilometres.
}
\details{
This function multiples the distance in miles by 1.60934 to get the distance in kilometres.

This function would be very helpful when using the package `nycflights13` or when you want to convert miles to kilometres.
}
\examples{
milestokm(1)
milestokm(250)
}
