% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_island.R
\name{plot_island}
\alias{plot_island}
\title{plot_island Draw a scatter plot for the island.}
\usage{
plot_island(data, island)
}
\arguments{
\item{data}{Optional: the datatset used to draw.}

\item{island}{island's name}
}
\value{
A ggplot picture
}
\description{
plot_island Draw a scatter plot for the island.
}
\examples{
\dontrun{
  plot_island(data = palmerpenguins::penguins,
              island = "Dream")
}
}
