% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_island.R
\name{table_island}
\alias{table_island}
\title{get table by island}
\usage{
table_island(data, island)
}
\arguments{
\item{data}{Optional: the dataset to be used}

\item{island}{which island to glance}
}
\value{
table
}
\description{
get table by island
}
\examples{
penguins <- palmerpenguins::penguins
table_island(data = penguins, island = "Torgersen")
table_island(data = penguins, island = "Dream")
}
